pyproject.toml
[project]
name = "evening-telegram"
version = "0.1.0"
description = "A Python application that reads Telegram channels and generates a newspaper-style digest using LLMs"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Evening Telegram", email = "noreply@example.com"}
]

dependencies = [
    "telethon>=1.34",
    "python-telegram-bot>=20",
    "openai>=1.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "click>=8.0",
    "jinja2>=3.0",
    "aiosmtplib>=3.0",
    "aiosqlite>=0.19",
    "rich>=13.0",
    "python-dateutil>=2.8",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
    "pytest-cov>=4.0",
    "ruff>=0.1",
    "mypy>=1.0",
]

[project.scripts]
evening-telegram = "evening_telegram.cli:main"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
